refactor: split pack checks facade
